requirements.txt
bcrypt==3.2.0
boto3==1.14.51
botocore==1.17.51
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
docutils==0.15.2
elasticsearch==7.9.1
elasticsearch-dsl==7.2.1
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
neo4j==4.1.1
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.1
paramiko==2.7.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scp==0.13.2
six==1.15.0
tqdm==4.48.2
uritemplate==3.0.1
urllib3==1.25.10
